requirements.txt
attrs==19.1.0
autopep8==0.8
aws-cdk.assets==1.5.0
aws-cdk.aws-apigateway==1.5.0
aws-cdk.aws-certificatemanager==1.5.0
aws-cdk.aws-cloudformation==1.5.0
aws-cdk.aws-cloudfront==1.5.0
aws-cdk.aws-cloudwatch==1.5.0
aws-cdk.aws-ec2==1.5.0
aws-cdk.aws-elasticloadbalancingv2==1.5.0
aws-cdk.aws-events==1.5.0
aws-cdk.aws-events-targets==1.5.0
aws-cdk.aws-iam==1.5.0
aws-cdk.aws-kms==1.5.0
aws-cdk.aws-lambda==1.5.0
aws-cdk.aws-logs==1.5.0
aws-cdk.aws-route53==1.5.0
aws-cdk.aws-route53-targets==1.5.0
aws-cdk.aws-s3==1.5.0
aws-cdk.aws-s3-assets==1.5.0
aws-cdk.aws-s3-deployment==1.5.0
aws-cdk.aws-sns==1.5.0
aws-cdk.aws-sqs==1.5.0
aws-cdk.aws-ssm==1.5.0
aws-cdk.core==1.5.0
aws-cdk.cx-api==1.5.0
aws-cdk.region-info==1.5.0
boto3==1.9.206
cattrs==0.9.0
jsii==0.14.3
mypy-extensions==0.4.1
publication==0.0.3
pycodestyle==2.5.0
python-dateutil==2.8.0
six==1.12.0
typing-extensions==3.7.4
